chat/requirements.txt
#flask
#gunicorn
#flask-cors
#apscheduler
#qdrant-client
#requests
#beautifulsoup4
#tiktoken
#python-docx
#python-pptx
#openpyxl
#pypdf
#pdfminer.six
#google-chat-python-client
#google-auth
#google-auth-httplib2
#msal

fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.4
qdrant-client==1.15.1
requests==2.32.5
apscheduler
beautifulsoup4
google-api-python-client
google-auth
google-auth-httplib2
msal
